# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_api_handlers.TriggerJobRequest:
    properties:
      job_id:
        type: string
    required:
    - job_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@stackservice.com
    name: API Support
    url: http://www.stackservice.com/support
  description: GenZ Web3 Multi-Chain Investment Platform API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Stack Service API
  version: "1.0"
paths:
...
  /api/v1/admin/workers/restart:
    post:
      description: Stops and restarts the wallet provisioning scheduler
      produces:
      - application/json
      responses:
        "200":
          description: Scheduler restarted
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Restart worker scheduler
      tags:
      - admin
  /api/v1/admin/workers/status:
    get:
      description: Returns the current status of the wallet provisioning worker and
        scheduler
      produces:
      - application/json
      responses:
        "200":
          description: Worker status
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get worker status
      tags:
      - admin
  /api/v1/admin/workers/trigger:
    post:
      consumes:
      - application/json
      description: Manually triggers processing of a specific wallet provisioning
        job
      parameters:
      - description: Job ID to process
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.TriggerJobRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Job processing triggered
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Trigger job processing
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"